Build maintenance: Fix a failed test by calling a internal method
The test failed due to a expired SMD certificate. The stacktrace was:

	godaddy.registry.jtoolkit2.se.tmch.exception.TmchInvalidCertificateException:
	Invalid Certificate provided in SignedMarkData. Certificate of serial number '7'
	and DN 'EMAILADDRESS=revoked@..., CN=revoked, OU=Dev, ..., C=AU'
	and signer DN 'EMAILADDRESS=..., CN=..., OU=Dev, ..., C=AU'
	valid from '2013-02-14T00:29:43Z' and valid to '2023-02-12T00:29:43Z'

Temporarily change the test to utilise Java reflection to call the other
private methods, which takes a second parameter of Date. So this test case can
pass in a date within the valid date range.

Will keep this test case like this, until the set of SMD files are re-generated.